smaller refactor to hmac
